requirements.txt
SQLAlchemy~=2.0.30
aiogram~=3.7.0
python-dotenv~=1.0.1
openai~=1.31.0
asyncpg~=0.29.0